Acceleration in float and the function drawCircle added

+++ Engine/Dude.h
@@ -6,17 +6,17 @@
 class Dude {
 
 public:
-	Dude(int in_x, int in_y);
-	int GetX() const;
-	int GetY() const;
+	Dude(float in_x, float in_y);
+	float GetX() const;
+	float GetY() const;
 	void checkBorders();
 	void Draw(Graphics& gfx) const;
 	void MovePressCheck(Keyboard& kbd);
 
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	static constexpr float width = 20.0f;
+	static constexpr float height = 20.0f;
 	
 private:
-	int x;
-	int y;
+	float x;
+	float y;
 };
